Keep the `State` schema and `result` stub in `prefect.orion.schemas`
