+++ src/prefect/states.py
@@ -4,24 +4,13 @@ import traceback
 import warnings
 from collections import Counter
 from types import TracebackType
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    Dict,
-    Generic,
-    Iterable,
-    Optional,
-    Type,
-    TypeVar,
-    Union,
-    overload,
-)
+from typing import TYPE_CHECKING, Any, Dict, Iterable, Optional, Type, TypeVar
 
 import anyio
 import httpx
-from pydantic import Field
 from typing_extensions import TypeGuard
 
+from prefect.client.schemas import State
 from prefect.deprecated.data_documents import (
     DataDocument,
     result_from_state_with_data_document,
@@ -41,145 +30,41 @@ if TYPE_CHECKING:
 R = TypeVar("R")
 
 
-class State(schemas.states.State.subclass(exclude_fields=["data"]), Generic[R]):
+def get_state_result(
+    state: State, raise_on_failure: bool = True, fetch: Optional[bool] = None
+):
     """
-    The state of a run.
+    Get the result from a state.
+
+    Implementation of `State.result()`
     """
+    if fetch is None and (
+        PREFECT_ASYNC_FETCH_STATE_RESULT or not in_async_main_thread()
+    ):
+        # Fetch defaults to `True` for sync users or async users who have opted in
+        fetch = True
 
-    data: Union["BaseResult[R]", "DataDocument[R]", Any] = Field(
-        default=None,
-    )
-
-    @overload
-    def result(self: "State[R]", raise_on_failure: bool = True) -> R:
-        ...
-
-    @overload
-    def result(self: "State[R]", raise_on_failure: bool = False) -> Union[R, Exception]:
-        ...
-
-    def result(self, raise_on_failure: bool = True, fetch: Optional[bool] = None):
-        """
-        Retrieve the result
-
-        Args:
-            raise_on_failure: a boolean specifying whether to raise an exception
-                if the state is of type `FAILED` and the underlying data is an exception
-            fetch: a boolean specifying whether to resolve references to persisted
-                results into data. For synchronous users, this defaults to `True`.
-                For asynchronous users, this defaults to `False` for backwards
-                compatibility.
-
-        Raises:
-            TypeError: If the state is failed but the result is not an exception.
-
-        Returns:
-            The result of the run
-
-        Examples:
-            >>> from prefect import flow, task
-            >>> @task
-            >>> def my_task(x):
-            >>>     return x
-
-            Get the result from a task future in a flow
-
-            >>> @flow
-            >>> def my_flow():
-            >>>     future = my_task("hello")
-            >>>     state = future.wait()
-            >>>     result = state.result()
-            >>>     print(result)
-            >>> my_flow()
-            hello
-
-            Get the result from a flow state
-
-            >>> @flow
-            >>> def my_flow():
-            >>>     return "hello"
-            >>> my_flow(return_state=True).result()
-            hello
-
-            Get the result from a failed state
-
-            >>> @flow
-            >>> def my_flow():
-            >>>     raise ValueError("oh no!")
-            >>> state = my_flow(return_state=True)  # Error is wrapped in FAILED state
-            >>> state.result()  # Raises `ValueError`
-
-            Get the result from a failed state without erroring
-
-            >>> @flow
-            >>> def my_flow():
-            >>>     raise ValueError("oh no!")
-            >>> state = my_flow(return_state=True)
-            >>> result = state.result(raise_on_failure=False)
-            >>> print(result)
-            ValueError("oh no!")
-
-
-            Get the result from a flow state in an async context
-
-            >>> @flow
-            >>> async def my_flow():
-            >>>     return "hello"
-            >>> state = await my_flow(return_state=True)
-            >>> await state.result()
-            hello
-        """
-
-        if fetch is None and (
-            PREFECT_ASYNC_FETCH_STATE_RESULT or not in_async_main_thread()
-        ):
-            # Fetch defaults to `True` for sync users or async users who have opted in
-            fetch = True
-
-        if not fetch:
-            from prefect.deprecated.data_documents import (
-                DataDocument,
-                result_from_state_with_data_document,
+    if not fetch:
+        if fetch is None and in_async_main_thread():
+            warnings.warn(
+                "State.result() was called from an async context but not awaited. "
+                "This method will be updated to return a coroutine by default in "
+                "the future. Pass `fetch=True` and `await` the call to get rid of "
+                "this warning.",
+                DeprecationWarning,
+                stacklevel=2,
             )
-
-            if fetch is None and in_async_main_thread():
-                warnings.warn(
-                    "State.result() was called from an async context but not awaited. "
-                    "This method will be updated to return a coroutine by default in "
-                    "the future. Pass `fetch=True` and `await` the call to get rid of "
-                    "this warning.",
-                    DeprecationWarning,
-                    stacklevel=2,
-                )
-            # Backwards compatibility
-            if isinstance(self.data, DataDocument):
-                return result_from_state_with_data_document(
-                    self, raise_on_failure=raise_on_failure
-                )
-            else:
-                return self.data
+        # Backwards compatibility
+        if isinstance(state.data, DataDocument):
+            return result_from_state_with_data_document(
+                state, raise_on_failure=raise_on_failure
+            )
         else:
-            from prefect.states import get_state_result
+            return state.data
+    else:
+        from prefect.states import fetch_state_result
 
-            return get_state_result(self, raise_on_failure=raise_on_failure)
-
-    def to_state_create(self) -> schemas.actions.StateCreate:
-        """
-        Convert this state to a `StateCreate` type which can be used to set the state of
-        a run in the API.
-
-        This method will drop this state's `data` if it is not a result type. Only
-        results should be sent to the API. Other data is only available locally.
-        """
-        from prefect.results import BaseResult
-
-        return schemas.actions.StateCreate(
-            type=self.type,
-            name=self.name,
-            message=self.message,
-            data=self.data if isinstance(self.data, BaseResult) else None,
-            state_details=self.state_details,
-        )
+        return fetch_state_result(state, raise_on_failure=raise_on_failure)
 
 
 def Scheduled(
@@ -438,7 +323,7 @@ async def return_value_to_state(result: R, result_factory: ResultFactory) -> Sta
 
 
 @sync_compatible
-async def get_state_result(state, raise_on_failure: bool = True) -> Any:
+async def fetch_state_result(state, raise_on_failure: bool = True) -> Any:
     """
     Get the result from a state.
     """
